+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to
 - The canonical format (`-Tcanon`) produces output that is stable across
   multiple invocations. #2193
 
+- Cylinder shapes now honor the `orientation` attribute. #2244
+
 ## [16.1.0] – 2026-09-03
 
 ### Added